// generated/src/app/api-models/id-card-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IdCardSetup {
  seqIdprtId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  requestType: string;
  action: string;
  seqNarrativeId: number;
  idCardTemplate: string;
  effectiveDateFrom: Date;
  effectiveDateThru: Date;
  orderType: string;
  reprintRequest: string;
  reprintDate: Date;
  status: string;
  comments: string;
  template: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  inProcess: string;
}